Refactor Toolkit TLS transport
